Implement bingo line difficulty
Users can now separate their bingo lines into easy, medium or hard
difficulties, and specify how many of each will appear on every grid.

This also necessitated some refactor of the types and the output logic
to keep things clean.

I also tried updating material-ui to squash an error in Brave:
`findDOMNode is deprecated in StrictMode`, which appears when the
Backdrop is opened on the render step. No dice. Should get fixed in MUI v5.

=====
Squashed commit of the following:

commit 967d1f7f5fcffe25409c35b6629af46ec096edc3
Author: Derek Surguy <Surguy.Derek@gmail.com>
Date:   Sat Sep 11 14:52:52 2021 -0500

    Add missing font sizes and format progress dialog

commit d9dd8a3e4588b6968798fb3cb096fe6048f912ac
Author: Derek Surguy <Surguy.Derek@gmail.com>
Date:   Sat Sep 11 14:47:35 2021 -0500

    Update material-UI to latest v4

commit b1bc5a6931cba3e511b6de9bea2c1303732b6b26
Author: Derek Surguy <Surguy.Derek@gmail.com>
Date:   Sat Sep 11 14:42:55 2021 -0500

    Split lines into difficulties